requirements.txt
mediapipe==0.8.9.1
numpy==1.22.3
opencv_python_headless==4.5.4.60
Pillow==9.1.1
pyserial==3.5
streamlit==1.10.0
tensorflow==2.8.0
